pyproject.toml
[tool.poetry]
name = "rag-bot"
version = "0.1.0"
description = ""
authors = ["vesalius512 <vladimiravramenko99@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain-community = "^0.4.1"
langchain-chroma = "^1.0.0"
langchain-text-splitters = "^1.0.0"
datasets = "^4.4.1"
python-dotenv = "^1.2.1"
sentence-transformers = "^5.1.2"
langchain = "^1.1.0"
langchain-huggingface = "^1.1.0"
openai = "^2.8.1"
tqdm = "^4.67.1"
praw = "^7.8.1"
orjson = "^3.11.5"
xdk = "^0.5.0"
aiohttp = "^3.13.2"
requests = "^2.32.5"
pre-commit = "^4.5.1"
ruff = "^0.14.10"
dynaconf = "^3.2.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
